This is a little overhaul of the directory structure.
Now include directories are placed inside their corresponding
source directories so that the CMakeLists.txt only needs a to include
the relative path "include".

Also Power was moved inside Model.
In addition I added some work on the model, which is bad from a commit
standpoint and I will do better in the future.