Minor changes in CMakeLists.txt
The project now is listed as a C and CXX project, because the check for inter
process optimization (a.k.a. link time optimization) fails otherwise.

Also the warning level of Visual Studio is reduced to three because the linting
otherwise clutter the output. Maybe there is a way to suppress messages and
only display warnings with Visual Studio. In that case, that should be used
instead.