Fixed Link time optimization and enabled building with clang and libc++
The project now is listed as a C and CXX project, because the check for inter
process optimization (a.k.a. link time optimization) fails otherwise.

Also the warning level of Visual Studio is reduced to three because the linting
otherwise clutters the output. Maybe there is a way to suppress messages and
only display warnings with Visual Studio. In that case, that should be used
instead.

* Added an option to build with libc++ when using clang

The option is called BUILD_WITH_LIBCPP and is OFF per default.
It can be invoked via cmake -DBUILD_WITH_LIBCPP=on ...
When using it with any other compiler it has no effect.


Co-authored-by: Eike <e.fokken+git@posteo.de>