Clearer status message for ipo and stricter debug build
The ipo message now states, that ipo is supported instead of activated,
because for debug builds it is not activated.

The warning level of msvc is now again W4 and warnings are treated as errors.
Probably we should check on all this clutter in the output but it has no priority.