faster Equationcomponent
I moved the vector times and values into a new struct "Valuecontainer".
Equationcomponent holds a unique pointer to the struct, to decrease
the size. As times and values isn't in "hot code", this is a performance
benefit (eyeball benchmark with gnu time).
The performance benefit would likely turn into a loss, if used on
problems where Newtons method needs very little iterations in every step.
I haven't tested this, but guess that this may turn up, if the average
number of iterations is less than 5.
on average.