The component schemas are now stricter.
They now actually confirm that the json has the right format for
the constructors to take them.
To still make a useful topology schema, the function get_topology_schema
of Component_factory calls the newly introduced function
relax_schema, which takes a number of keys that are remove from the
"required" list in the schema.