ImplicitOptimizer:
This now holds a uniqe_ptr to a cache. This makes testing easier.
Also dropped the requirement, that the constraints share the same last time
with the states. Instead, we now check, when constraints end.

Timeevolver now has a factory to produce a unique pointer instance.

ControlStateCache now holds a unique pointer to Timeevolver and doesn't
have a problem reference anymore. Instead the problem reference must be
supplied to its method as an argument.