removed set_coefficient in matrix handler.
Instead the Coeffrefhandler now explicitly zeros out all coefficients on
construction of the handler. This means that we have no more
handler that can update an existing matrix. But we don't need that
functionality and in this way, noone ever has to think about
whether they should call set_coefficient or add_to_coefficient.

This will also enable a intuitive implementation of penalty functions.