Penalty objectives (#328)
* added penalty



* removed set_coefficient in matrix handler.

Instead the Coeffrefhandler now explicitly zeros out all coefficients on
construction of the handler. This means that we have no more
handler that can update an existing matrix. But we don't need that
functionality and in this way, noone ever has to think about
whether they should call set_coefficient or add_to_coefficient.

This will also enable a intuitive implementation of penalty functions.

* added a test to check that coeffrefhandler constructor sets matrix to zero.

* added functions to evaluate the penalty

* Now also penalty and cost are provided in the terminal output.

* Also state the output directory on terminal output

* changed to macos latest in actions as 10.15 is depreciated.

* clang format