disable warning against missing defaults in switch statements in tests dir
googletest defines some macros that have switch statements without
explicit defaults. This is (hopefully) fine because they (hopefully) make
sure that the default is never reached. But as it happens in a macro,
the issue is found when compiling grazer sources and so is triggered
even when googletest is linked as a system library. So we ignore this
issue in test files but not in the actual source files.