remove the macos pipeline.
Unfortunately there seem to be some issues with macos on arm.
When run on the runner, ctest says there are "ILLEGAL" exceptions,
which apparently means some wrong instructions were used for the cpu.
Possibly our configuration implicitly cross-compiled to non-arm, but
I have no mac and cannot debug this. Hopefully, running grazer on a mac
still runs fine, but we don't test that, anymore.