disable a warning in tests and remove macos runner (#446)
* disable warning against missing defaults in switch statements in tests dir

googletest defines some macros that have switch statements without
explicit defaults. This is (hopefully) fine because they (hopefully) make
sure that the default is never reached. But as it happens in a macro,
the issue is found when compiling grazer sources and so is triggered
even when googletest is linked as a system library. So we ignore this
issue in test files but not in the actual source files.

* remove the macos pipeline.

Unfortunately there seem to be some issues with macos on arm.
When run on the runner, ctest says there are "ILLEGAL" exceptions,
which apparently means some wrong instructions were used for the cpu.
Possibly our configuration implicitly cross-compiled to non-arm, but
I have no mac and cannot debug this. Hopefully, running grazer on a mac
still runs fine, but we don't test that, anymore.